fix(compat): preemptive review hardening
- Extract createStaticRepositoryServer to tests/support/ to eliminate
  code duplication between browser-ready-distribution and compat tests.
  Shared helper includes .svg MIME type and uses a named getRequestPath
  function consistently.
- Replace node_modules/.bin/ paths with 'npm exec --' in build script,
  test file, and README manual commands (cross-platform portability).
- Guard server in afterAll: skip close if server never started.
- Add examples/compat/*/dist/ to .gitignore so build artifacts are
  not shown as untracked files.
- Wrap execSync in build-compat-examples.mjs with try/catch for
  clear error label and clean exit on build failure.
