+++ .gitignore
@@ -37,6 +37,9 @@ test-screenshot.png
 Thumbs.db
 *.bak
 
+# Compat example build outputs
+examples/compat/*/dist/
+
 # Spurious tsc output in src/ (outDir should be dist/)
 src/**/*.js
 src/**/*.js.map
